Set homonym parent to replacement name parent in import
Not 100% sure this is the right thing to do, but I haven't found a reason why it shouldn't work.

It's needed in the case where the homonym's parent is a synonym originally at species level, but later moved to subspecies to match the valid name's rank. At that point, the homonym's parent would be a subspecies, which is impossible. To resolve this, we can set the parent of the homonym to the parent of the valid replacing name.

Ideally we'd have a user-controllable setting for this behavior, but until then the behavior will be enabled by default.