Make the entire task hub card clickable
Previously the link-clickable portion of the card was only as wide as the task title and description, which for some cards made much of the card unclickable.

The other fix here is to prevent the task card link from extending outside the bounds of the card border: previously that was happening when the 30px task-information left margin was pushing its 100% width (which doesn't account for margin) outside the task card border when the link was long. Now with `flex-grow: 1` on the link, the link's children grow to the size of the link, so an explicit width on task-information isn't needed.