#1691 Restrict deletion of keys that have a redirect to them
NOTE: testing suggests that in the case of destroying a key that redirects to itself, the result of the added dependent: may depend on which order the child nodes get destroyed: if the child node with !redirect_id.nil? gets destroyed before the node it redirects to then destruction of the key will succeed, otherwise it will not. That's fine from a destruction standpoint (the thing doing the referring is gone so it's fine that the referred goes away), though it may be confusing for observant users - in any case I don't know offhand how to fix it.