#1691 Fix a browser history/turbolinks/spinner bug involving new_lead
STR:
Start at /leads/1
Click the Edit link.
Click the browser's history back button.
Click the browser's history forward button.

See 'domElement is nil' or similar error from init() in Spinner.vue.
In this case domElement is a specified div target of a CornerSpinner. init is being called in mount(). Assigning to domElement using getElementbyId for the div's id fails, which leads to an exception and a red overlay. When you close the overlay you see the page being displayed is still /leads/1 even though the browser url is the edit link.

So ... wild guess ... vue constructs the new vue app dom for edit and thinks everything is ready to go, but spinner's getElementById call happens in mount() before turbolinks has actually replaced the old content (/leads/1) with the new edit content, so domElement is never found.

I wonder if this is going to pop up again with other links.