#1691 Close the meta layout once editing on the key starts
Also display the key title in the meta layout header when that layout is not expanded.

Right now the state of the meta layout expansion is preverved on browser back and forward through editing pages, but not if the page is reloaded or on browser back/forward from a non-editing page to an editing page.