#1691 In new_lead, previous and future links should point to parent
A bit confusing, but when you click on some text from a couplet in past or future, you want to load so that that text is on left or right, which means you should loadKey the parent of the lead with that text.

Also a bit confusing (to me): previous displays the texts that have already been selected, along with the couplet number of those texts. The couplet number of a given lead text is stored on the /parent/ of the lead with the given text, it looks like this:

[couplet number]     or     [couplet number]
       /\                          /\
   text  *                        *  text

The raised (superscript) couplet number in the display of previous couplets adopted here is meant to indicate that relationship (though it doesn't indicate left or right).