#1691 Remove autocomplete_exact_id
I think autocomplete_exact_id is likely to be confused with search results from autocomplete exact couplet number, and used much less often. The one exception I can think of is if you've duplicated a key and the only way to disambiguate between duplicated leads is via id; you'll just have to edit the key directly in that case.