Only set spinner top and left if position is 'fixed'
Ironically this restores the behavior prior to my adding 'px' units to the top and left values of `copyCSS` in an earlier commit in the Leads work.

Unless I'm mistaken the CornerSpinner I created in Leads is currently the only spinner that specifies a target prop on a spinner, in all other cases in use the target of the spinner is the parent node of the spinner. In that case setting top and left of the spinner with absolute positioning to be top and left of the containing element is probably not going to do what you want - see how it affects the use of spinners to hide content in /tasks/nomenclature/new_taxon_name, for example.

Even in the case that the spinner's position is specified as fixed, there's still the question of what the top and left values of the target element mean - if the target element is also position fixed then all is well, but if not then the spinner position probably will not end up coinciding with the target element's position.