#3881 Stub out the new Leads Hub task
TODOs:
* Include metadata section like source hub?
* Support sorting by column in the keys list (might depend on if we need to page that list)
* Make 'last updated' text nicer (maybe like /leads/list?)
* Make 'Is public' an editable checkbox in the keys list? (Is that something other tasks do?)
* Implement
  * show entire key on one page, html
  * show entire key on one page, for printing
  * download markdown and/or latex versions of a key

Right now it seems like the two 'entire key' options should go on the radial navigator for a lead, but the download(s) should not, but I'm not sure about that. If everything can go on a radial then there's no need for the Options button, but I'm thinking if at least one of those wouldn't go on the navigator then we might as well list all of them as Options along with the download(s) option.