Enable sorting of columns in Key Hub key list
Note that since there's only one `ascending` ref for all columns (which tracks whether the next column sort will be ascending or descending), clicking on one column and then a different column will sort the columns differently. That seems fine since we don't have indicators on the columns indicating which order the column is in.