Move Is Public to the last column of keys hub
One side-note on a css change here: prior to moving the Is Public column the first 3 columns were Title, # Couplets, and Is Public. The first has width: 40% and the last two had width: 4em; the otus column spanned those three. Now that the Is Public column is moved and the otus column only spans the first two, when the otus column has content that requires wrapping, the # Couplets column grows past its specified width, from say 52px to 72px - this doesn't seem to depend on the exact contents of the otus column, only that their content requires more than some (unknown) width. Not a big deal if that's all the more it ever grows, but it is noticeable when it happens, and I don't know why it happens so I don't know that it won't ever grow more. If you add a hack column with set width after the # Couplets column then all works as expected again, all of the other width/min-width/max-width/table-layout:fixed attempts I made failed in some way. Shrug (but if you know why this happens please let me know!).

The current solution of changing the width: 4em; columns to width: 1%; seems to work as expected, the only issue I have is I would like the # Couplets column to be a little wider (with the 1% width it shrinks to the narrowest word in its header, which is "Couplet"), but I think it's fine.