Correct order of roots and update based on new understandings
My new understandings:
* If you use a symbol in an `.order` like `M.query.order(:text)`, the order applied to your sql statement will be `T.text`. If you use a string literal like `M.query.order('text')` then the order applied to your sql statement will be `text` (not `T.text`).

If your query was a join that returned the text column from not `T` but a joined table of the same type as `T`, then `.order(:text)` will be wrong because it will order by text from the first table of the join. If you do `.order('text')` that will give the right result because now it looks at the returned columns to find one with an attribute named `text` (instead of explicitly specifying `T.text`). Even more explicit is to do `.order('leads_updated_at.text')` where `leads_updated_at` is the alias of the second table in my join from which I'm pulling leads. That's what I did here.

* If you have a join like `q = T.joins('JOIN U on T.id = U.id').select('U.*')`, where both `T` and `U` have a column named `text` (say if `T` and `U` are the same model), then `q.pluck(:text)` will pluck text from `T`, not `U`, because the `joins` was called on `T` (not `U`). This time `q.pluck('text')` doesn't help, it's still added to the pluck query as `T.text`. `q.pluck('U.text')` does work, I guess because it has a '.' in it.

But there's no way anyone using `roots_with_data` would know to use pluck that way if I were to return the raw query, so instead I've changed it to return `q.to_a`. (Maybe returning the query wasn't a good idea in the first place.)

* When you do a join like `q = T.joins('JOIN U on T.id = U.id').select('U.*)`, the resulting rows are:
  * of type ActiveRecord for type `T` (because that's what you called the `joins` on)
  * having the attributes of `U` (and only the attributes of `U`)
Unless T = U, as we have here, that's sort of nonsense. So in a sense, in this case where I *have* to pull results of the join from the second table `U` (the `T` part of the resulting rows is useless), it's maybe just luck that `U` and `T` are both Leads, and so when I select `U.*` I am creating an ActiveRecord of type Lead with the correct data from `U`.