#2029 Support filtering documents by extension in the Import Nexus task
I tried as much as I could to write things so they could be reused in a Filter Documents task (in case that ever comes to be and if the facet I've written is desired there). In that event maybe we'll want to filter by document type as well, and maybe the user should be able to enter an extension and/or type.