#2029 Add option to name the new matrix
As I noted at the defineModel() for the options ref, and as the vue docs point out: if you define your model ref default in defineModel, the default is not passed to the parent - so if your options, like matrix name, are never touched, then you wind up passing an empty options hash to the controller and it raises on an empty required param object.

The solution is simply to assign the default outside of defineModel().