#2029 Move preview code to nexus vendor helper module
I'm still not sure if this is the right place for this code or not, but having it all in the controller doesn't seem right. Maybe lib/vendor/nexus_parser.rb would be better?

The other motivation here is that the preview code and the import code share some functionality - the import is run from a background job, so the shared code needs to be somewhere both the controller and the job have access to it (maybe the data created by the shared parts could be injected).

This leaves the import path temporarily broken.