#2029 Fix bugs related to matrix_name option, add specs
The matrix_name option is processed in the controller code so that we can report any errors to the user right away instead of via the background job, where they'll never see the error unless they ask an admin. Unfortunately that means all other options are tested in the helper spec since those options are processed in the helper code.