#2029 Make extensions data an initializer constant
In my case I still need to refer to particular extension groups by group name strings, for example in vue I need to select the extension group with name 'nexus'. But at least now the group names are (can be) constant over time even if extensions are changed.

I'm also not sure about using the group name '' to mean any extension. That fits with filtering (blank means nil facet, which translates to any extension), but it's a little confusing in vue-land where it may not be obvious what '' for a group extension group name means. Maybe group name 'any_extension' would be better in that case.