#2029 Compute ExtensionGroups in Facet instead of document selector
The situation is:

DocumentSelector
    FilterDocument
        FacetFileExtension

Both DocumentSelector and FacetFileExtension need the ExtensionGroups data loaded from a Document endpoint - it was loaded in DocumentSelector and then prop-ed down to the Facet, but I think if the Facet is ever used in a Document filter then it should load its own data. So now we param it up the chain instead.

I think this way is actually cleaner as well.