#2029 Don't unique the otus and characters lists when previewing
Uniquing was a carry over from a previous algorithm that's no longer used, and causes the otu/character numbers displayed in the ui to be smaller than those in the nexus file when there are repeats, which confused me.

It might be useful to warn of repeats in this case though.