#2029 Bundle citation options
The one catch here is that if you send an empty options hash param to the controller, rails doesn't include a parameter for it, and then your params.require(:options) fails (it would also fail if options *was* sent as {}...).

My old workaround was to set the default values for options in vue (where really to solve the empty options hash issue I would have only needed to set one of them) - which I've removed here in favor of (re)creating an empty hash in the controller and then permitting it as the options hash. I like that better - I no longer need to list out all of my options in vue so that I can assign them defaults (the matching and the citation params are now encapsulated in their own components), and the new rails workaround just acknowledges that rails discards #empty? params.