Export TW gap charstates as the special nexus gap state
Character states in TW must have a non-empty name, but nexus gap states are (typically) specified in a matrix by a special (global) gap symbol that is not a character state and has no name.

With the previous export, where we were adding 'gap' as a character state on descriptors which used a gap state, and then put a dash in the matrix table as the label for that state, Mesquite was:
* interpreting the '-' in the matrix table as a nexus (global) gap state, even though we weren't specifying a nexus gap symbol [also note that '-' is not in the list of symbols that we export, though I doubt that's why Mesquite interprets '-' as gap], and
* also keeping the 'gap' character state for descriptors that used it, which messed up (sequential) labelling of character states for descriptors and led to incorrect interpretation of the matrix data.

We just assume here that the gap state, if used at all, has symbol '-', though that's not required by TW or nexus.