Compute Symbols list on nexus export
The assumption is that TW non-gap states are numbered sequentially starting at 0 - that's the nexus assumption regarding labelling of character states. It would still be possible to export if that wasn't the case, but it would require adding extra '_' filler character states or adjusting numbering in the Matrix on export (or maybe there's a nexus command that can do similar?).

Nexus does allow matrix entries of, e.g., 9 when there are only 4 listed states for that descriptor, but TW does not allow that (matrix entries come from existant character state labels), so we don't need to run through the matrix as well here, i.e. the highest matrix entry comes from an existing character state, and we're already counting those here (under the previous assumption of course!).