Add a User#preferences accessor
Thanks to Hernán for the suggestion and code.

For when preferences are empty, in this case because the project was imported via psql and so avoided the before_save check that normally would fill any unset preferences.

Empty prefs lead to exceptions on things like preferences.layout reads, which assume default preferences exist.

Note that in general the preferences returned here won't be saved to the database.