#1954 Clarify `select_geography_sql` and st_distance vs st_distance_spheroid
Apologies for getting somewhat opinionated here, I hope I'm not too far off...

The name and the description of `select_geography_sql` (as well as comparison with `select_geometry_sql`) suggest that the return value should be of type `geography` (it was previously being cast to type geometry).

But then I'm confused about `st_distance` vs `st_distance_spheroid`:
* in `st_distance_spheroid` we cast geography types to geometry types, but then do distance calculations on a spheroid.
* in `st_distance`, prior to the `select_geography_sql` change here, we were calculating planar "distance" between two geometry arguments and then converting it to meters - but the use of `select_geography_sql` in that function suggests that what was intended was to calculate `ST_Distance` between two geography points, which is a spheroidal calculation.

So I would expect the same answer in both cases, which is what I'm getting here. As to which is faster or slower, I can't say, but `st_distance` seems more natural to me, though `st_distance_spheroid` is the one that's used in other places besides specs.

If we really wanted a faster, less accurate calculation I think it would be `st_distance` between two geometry types (and then converting to meters? I'm unclear there) - should one of these be converted to that?