#1954 Adjust GeographicItem#orientations for geography type
Still not sure what the right way to expose the underlying shape of the geography column (is it point, polygon, etc). Right now there are lots of competing types and geo_types and object_types and multiple versions of those names; hopefully when it's just geography that can be simplified.