#1954 Rewrite GeographicItem query in CollectingEvent to not use GEOMETRY_SQL
Don't want to be using GEOMETRY_SQL outside of GeographicItem.

I may well be missing something, but in regards to the comment about not wanting to load the entire GeographicItem, my thinking is that the size of the shape dwarfs the rest of a geoitem in general, and the rewrite here fetches that shape as wkb instead of as geojson (which i think would be larger?).

Also I copied the snippet here from the same function in Georeference :)