#1954 Deprecate more GeographicItem methods
D'oh, I factored out st_distance_item_to_shape and then realized where I was using it was only used in specs.

My feeling has been that it's better to select the geography column of a fixed geo_item from the database (once, as a subquery) rather that sending its geo_object, which could be "very" large, over the network and then having the database read it into memory. Thoughts?