#1954 Remove/replace GeographicItem#rgeo_to_geo_json
I may just be confused here, but to me it looks like rgeo_to_geo_json was returning a json string, not a hash as the @return said. I think we want a hash for the geometry key of the geojson feature (as a hash) in to_geo_json_feature?

In that case why not just use to_geo_json for a geometry collection for the value of the geometry key? (Maybe it wasn't supported when this code was originally written?)