#1954 Change names of st_isvalid, st_isvalidreason, add specs
As I note in the comments, the way it's looking to me is that rails/rgeo/rails-rgeo-adapter is already testing is_valid as a rails validation on the geography column; if so then st_is_valid will always return true.

Also it seems odd that geographic_item.valid? is raising an (rgeo) error instead of just returning false