#1954 Create a DocumentSelector for gazetteer shapefile documents
Similar to the nexus import doc selector, but in this case you need to be able to select multiple files at once, since that's how shapefiles come (I don't think we want to allow zip file uploads??).

The other slightly awkward part here is that the docselector/smartselector lists are radios, assuming you just want to select one item; in this case each time you select a radio item it gets added to a list of items. I guess the interpretation here would be that the currently selected radio just indicates the most recent item added to the list...