Document unfortunate leaflet/GeoJSON.decode interaction
To expand on the added comment: When the leaflet map first comes up, it repeats the mercator projection multiple times in the horizontal direction. If you draw your shape in the middle projection you get what you'd expect, but if you draw on one of the other projection you get coordinates as if the projections are all combined into an endless longitude number line, so you get longitudes like 450, -720, etc.

```
RGeo::GeoJSON.decode(
  shape, json_parser: :json, geo_factory: Gis::FACTORY
 )
```

normalizes those coordinates for polygons, but not for points(!), so the check_point_limits throws an error in that case.