#1954 Update job import list functionality
One note on the job status trash button: the button deletes from the database but I've made it primary color with no confirmation here since the delete is of little consequence (?) and having to weed that table by hand is already a bit of a burden on the user. I have no strong feelings on that though.