#1954 Allow any GazetteerImport job to be deleted after one day
Prior to this a user couldn't delete a GzImport job until ended_at had been recorded - but if there's an error and it never gets recorded then the job's going to sit in the ui forever, or until someone complains to an admin.

The danger here is that an active job holds a a GzImport reference, which it assumes is valid and can be updated without checking validity, so if the user does delete the GzImport job before the job completes then all of the admins will be getting an email because of a GzImport not found exception.