#1954 Prevent empty required citations param error
I seem to keep running into this issue: you can't permit a parameter without requiring it, but it's not actually required in the ui (i.e./and/or the default blank values aren't passed to the server), so you get an error on the backend. In other words you can't pass empty hashes to express un-assignedness I guess.

(Unfortunately the fix here means I bleed the `cite_gzs` identifier into a parent component that doesn't otherwise need to know anything about it.)