#1954 Split anti-meridian crossing shapes across anti-mer. on New Gz save
My impression is that all GAs that "cross the anti-meridian" are actually split along the anti-meridian so that they don't in fact touch the anti-meridian. Given the lack of (perceived on my part, though possibly just my confusion) documentation/uniformity/even self-consistency of the gis libraries we use regarding their behavior on shapes that do cross the anti-meridian, that seems like a good idea here as well.

The strategy here is to
1) use GeographicItem.crosses_anti_meridian? to detect any input shapes that do cross the anti-meridian - for New Gz that's going to be geojson shapes drawn on leaflet and WKT shapes entered by hand
2) use the new GeographicItem.split_along_anti_meridian to decompose those shapes into shapes that come close to but don't touch the anti-meridian.

Points directly on the anti-meridian may be/come an issue as well (maybe depending on how their longitude is specified?)?