#1954 Document valid leaflet anti-meridian shapes becoming invalid in rgeo
Somebody who has dug into the undocumented anti-meridian properties of the various rgeo/postgis functions may be able to answer the question I raise in the comment about whether it's possible to write that shape as a single polygon in a `valid?` way.

I'm not sure, is this a reasonable way to document this?