#1954 Truncate coordinates display in New GZ
Mainly for the union-of-GAs-or-GZs case, where there may be hundreds of polygon rings and/or vertices per ring. I thought the large coordinate strings might have been slowing down page rendering of large shapes as well, but based solely on what I'm seeing playing around I'm now thinking it's more the rendering of the shapes themselves than the coordinate string (??).

If useful we could always add a button to expand the entire thing for copying (though in its current state I'm not sure what would be able to parse it) - and/or write it as valid WKT.