#1954 Add special case single point longitude normalizing for New GZ
A similar issue exists in current CE Georef creation if you:
* enter a wkt point with long outside -180 to 180
* enter a point by hand with the same
* add a point to leaflet that's outside of the central -180 to 180 part of the map

You get an error when the GI is saved saying the longitude is out of range.

In each of those scenarios, un-normalized longitudes get normalized by internal rgeo processing if you enter anything other than a single point, for example there will be no complaint with a WKT of 'POLYGON ((999 1, 995 2, 990 3))' and the polygon will be returned with normalized longitudes. In the New GZ case even the case of creating two points in leaflet outside the middle zone is handled correctly (since it becomes a multipoint, not a single point).

I don't know if this is technically an rgeo bug or if it's meant to be that way, in either case for the user's sake it seems worth handling so that they don't have to.