#1954 Abandon using one large transaction on GZ imports
Prior to this change the entire import ran in a transaction, the idea being that either all shapefiles shapes would be imported or none would be.

In my mind the issue was that if some succeeded and some failed then the user needed to sort out which ones hadn't been added, extract those from the shapefile (because you can no longer import the entire shapefile without generating duplicates of the previously successfully imported records), fix them somehow, and then attempt to reimport again, perhaps with different errors causing the process to repeat again.

In practice I now think that's overly pessimistic: we now make_valid all shapes, so the errors we can't handle are only those where geos raises on simply trying to read the shapefile shape and failing (even though rgeo::shapefilereader is configured with allow_unsafe: true to allow **rgeo** to pass on invalid shapes) so that we can't even get a hold of a shape to work with - so far in the wild those look like:
* shapes that cross the anti-meridian in a way that can be interpreted as causing a self-intersection (this depends on how lines between coordinates on either side of the anti-meridian are interpreted - the "long" way around across the meridian or the "short" way across the anti-meridian - which is a black hole of undocumented mess) [TODO if geos is checking for self-intersections on reads then what's the point of rgeo::shapefile's allow_unsafe: true option? am I missing something?]

* shapes that cross the +/-85.1... latitude pseudo-mercator boundary, where all "larger" latitudes seem to get clamped to that value and then become invalid polygons

In practice neither of those conditions should be very common. Also anyone who's able to fix errors of this type should also be capable of extracting the correct shapes from the original shapefile for resubmission without duplicates (n.b. `ogr2ogr -f "ESRI Shapefile" -makevalid fixed_shapefile.shp Ecoregions2017.shp` did **not** fix those issues and both rgeo and postgis reported each of the shapes flagged by rgeo as invalid as being valid).

Also regular users probably prefer to have most of the shapes imported while someone works on getting the rest fixed.

The other more pressing issue is that rails transactions seem to be quite memory hungry. On my system importing the (fixed) EcoRegions shapefile requires 11-12 gigabytes of memory.

I'm new to this but my chatgpt-supplied understanding of the issue is that:
a) when rails does a transaction that involves ActiveRecord saves, **every** ActiveRecord that's saved during the transaction is held until the end of the transaction, so memory usage grows and grows.
b) when an ActiveRecord is saved, all of its (non-associational) data is saved in memory - for us that's a GZ getting created/saved and its associated GI getting created/saved

EcoRegions has 847 records, let's call it 1,000, so that would be ~10,000,000,000(10G)/2,000 (a GZ + a GI for every record) = 5Mb per record. Record 753 is a multi-polygon with over a million vertices and many subshapes, so with transaction overhead, arrays, hashes, etc, maybe that's actually realistic???

I don't know what kind of memory most TW servers have, but for my home computer with 16Gb ram, a 12Gb memory job requires closing everything and switching to a tty, not cool.

So now the transaction is at the individual records level: if you're importing a record into multiple projects, that runs in a transaction so that you know either all projects got it or none did; if you're only importing into a single project (the default) then there are no transactions.

The errors mentioned above are now caught and reported in the ui as a concatenated string (which will quickly get clumsy if there are a lot of errors...).

The other added benefit is that we can now report per-record progress (well per 5 records) for long-running imports - I couldn't figure out a clean way to do that from inside a transaction before.