#1954 Support importing iso a2/a3 from shapefile in GZ import, add specs
This is the "right" way to test a dbf column's type (instead of testing the first record's type for that column), I just wish the api was a little cleaner. (Initially I wasn't sure what happens when a String column has an empty value: is the value returned for that column always "" or could it be nil? (My sample size one experiment coming from QGis was that it was "" and a quick look at the dbf code suggests the same, but I'm still not 100%.)