#1954 Test for and split anti_meridian_crossing shapes before doing make_valid
See the specs for reasoning (which, to be up front, are based largely on one shape demonstrating the issue).

This is actually already our behavior for New Gz processing, so that hasn't changed.

I thought we might be making a tradeoff here (gaining creating valid shapes that came from valid anti-meridian-crossing shapes that wouldn't have made it before, for invalid anti-meridian-crossing-shapes that would have made it before), but the example I was thinking of actually doesn't make it either way:

s = Gis::FACTORY.parse_wkt('POLYGON ((170 0, 200 10, 200 0, 170 10, 170 0))')
 => #<RGeo::Geographic::ProjectedPolygonImpl:0x6ec94 "POLYGON ((170.0 0.0 0.0, -160.0 10.0 0.0, -160.0 0.0 0.0, 170.0 10.0 0.0, 170.0 0.0 0.0))">
3.3.4 :003 > s.make_valid
 => #<RGeo::Geographic::ProjectedMultiPolygonImpl:0x70e18 "MULTIPOLYGON (((-160.0 10.0 0.0, 5.0 5.0 0.0, -160.0 0.0 0.0, -160.0 10.0 0.0)), ((170.0 10.0 0.0, 170.0 0.0 0.0, 5.0 5.0 0.0, 170.0 10.0 0.0)))"> <<--- it made the meridian-crossing version valid, as I should have guessed