#1954 Add support for shapefile .cpg for attribute encoding information
A few notes:
Attribute encoding can be stored in either a .cpg file (which seems to be the more modern approach) or as an `encoding=` value in the .dbf header. The dbf gem prefers the former when present, in which case the header value is ignored.

Dbf's assumed encoding for its stored string attributes in the rgeo::shapefile-reading case is:
`cpg_encoding || header.encoding || Encoding.default_external` where Encoding.default_external should be UTF-8

On the other hand, dbf always returns attributes in the following encoding:
`@encoding(i.e. the above value) ? value.force_encoding(@encoding).encode(Encoding.default_external, undef: :replace, invalid: :replace) : value`
so in other words in normal circumstances string attributes returned from dbf are encoded in UTF-8, which of course is what we want/need.