#1954 Finish documenting new understanding of anti_meridian behavior
This is all undocumented in the libraries as far as I can tell and I haven't looked at all of the library code, so the comments here are based on the time I've spent playing around with this stuff and puzzling over it. I think it might actually be correct.

I don't really like that the specs I've added have kind of turned into a story ... but they're really a documentation of the chain of reasoning that has led me to conclude that what I'm saying there is correct. If they're not convincing - or wrong - then things should be changed.

One other note: if s crosses the anti_meridian then it's impossible to use rgeo's make_valid if you're applying it to an rgeo object created by Gis::FACTORY (at least without first splitting s) since the factory is incapable of representing the longitudes of s in the way that rgeo interprets as crossing the anti-meridian and not the meridian. We would need to use a different factory if we didn't have ST_MakeValid.