#1954 Always keepcollapsed=false for ST_MakeValid
The make_valid process on a polygon, e.g., can result in line_strings or points - here we choose to drop any of those lower dimensional shapes created during the make_valid process. I'd rather remove this as an option so that we have consistent behavior across all shape creation (at least for now).

At least that's what I thought would happen, but on EcoRegions EcoID=753 - which is a multipolygon - ST_MakeValid still returns a geometry collection containing a multipolygon and a multilinestring...