#1954 Document postgis interpretation of edges for ::crosses_anti_meridian?
The xspecify of mysterious (to me) sometimes failures of crosses_anti_meridian? when one of the vertices has longitude (exactly, and only) 0 is disturbing, and I don't know exactly when it happens.

If it turns out to cause problems, one option would be to implement crosses_anti_meridian? with the following pseudo-code (written here for polygons):

def crosses_anti_meridian?(poly: p)
  prev = nil
  for each |vertex of exterior ring of p: v|
    if prev and line(prev, v) intersects anti-meridian
      # work around a false positive:
      if v.lon == 0 || prev.lon == 0
        continue
      end
      return true
    end
    prev = p
  end
  return false
end