#4056 Add delete_leaf action for deleting leads with no children
I initially wrote this to allow deleting a (mostly) arbitrary lead, even if it had children, but that seems a little risky, even with a big warning (in the sense that personally I think I would tend to say 'uh ... yeah, I think this is the subtree I want to delete (even though I haven't rechecked all of the nodes below this one)"). I'm open to doing the riskier operation though if others want to.