#4056 Implement 'Insert key'
The root of the inserted key is appended to the children of the lead inserted on.

If your inserted key is for a genus and the key you're inserting into is for its parent family, the inserted genus root text will then change to be whatever condition at that point of the key leads to the conclusion that you're in that genus. (If you already had a lead with that text you'd want to copy it to the newly inserted root and then delete the old lead.)