#4056 Rename 'option set' to 'couplet' in leads
I happened to check out this wikipedia article: https://en.wikipedia.org/wiki/Single-access_key which says "At each point in the decision process, multiple alternatives are offered, each leading to a result or a further choice. The alternatives are commonly called "leads", and the set of leads at a given point a "couplet". ... If the key has several choices it is described as polychotomous or polytomous. If the entire key consists of exactly two choices at each branching point, the key is called dichotomous."

I actually prefer 'couplet' over 'option set' (I think, only because it's the more well-known/recognized term), in spite of the jarring reaction I get to seeing couplet refer to something with other than two parts.