#4056 Replace 'return true on success, error string on failure' pattern
I've been finding myself going to that pattern to deal with functions that can raise multiple types of exceptions, but not feeling very good about returning either a bool or a string and having to carefully check for `== true` and not just present. The new pattern (storing errors in the model's error object and just returning true/false) seems like a good solution in the rails case.