Fix bug in CachedMap#latest_cached_map_item
I initially assumed this bug would cause #synced? to be wrong sometimes (namely when a new cmi is created after cached_map), but it turns out not to be the case:

```ruby
def synced?
  cached_map_items_reference_total == reference_count &&
    latest_cached_map_item.created_at <= created_at
end
```
The ref count check always fails (i.e. should always fail) when you add a new cached_map_item, so the created_at check - which would fail with the bug - isn't checked.

`latest_cached_map_item` is also used in the UI in `cached_map_sync_metadata` helper, so that's probably showing incorrect data in this case (when not synced because of a new cmi), but I haven't checked.